invalidconfig: update import path after core/def → store/def rename
