+++ comp/healthplatform/issues/invalidconfig/BUILD.bazel
@@ -11,8 +11,8 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//comp/core/config",
-        "//comp/healthplatform/core/def",
         "//comp/healthplatform/issues",
+        "//comp/healthplatform/store/def",
         "//pkg/config/lite",
         "//pkg/util/log",
         "@com_github_datadog_agent_payload_v5//healthplatform",
@@ -26,7 +26,7 @@ go_test(
     embed = [":invalidconfig"],
     gotags = ["test"],
     deps = [
-        "//comp/healthplatform/core/def",
+        "//comp/healthplatform/store/def",
         "//pkg/config/lite",
         "@com_github_stretchr_testify//assert",
         "@com_github_stretchr_testify//require",
