+++ comp/healthplatform/issues/invalidconfig/issue.go
@@ -16,7 +16,7 @@ import (
 	"github.com/DataDog/agent-payload/v5/healthplatform"
 	"google.golang.org/protobuf/types/known/structpb"
 
-	healthplatformdef "github.com/DataDog/datadog-agent/comp/healthplatform/core/def"
+	healthplatformdef "github.com/DataDog/datadog-agent/comp/healthplatform/store/def"
 )
 
 // Context keys produced by Check() and consumed by BuildIssue(). The keys are
